Renamed: component PickerAttributesReset.vue
