+++ apps/frontend/src/components/page-products/pickers/PickerAttributesReset.vue
